chore: fix precision

+++ src/main/java/ai/timefold/solver/benchmarks/competitive/Dataset.java
@@ -1,5 +1,6 @@
 package ai.timefold.solver.benchmarks.competitive;
 
+import java.math.BigDecimal;
 import java.nio.file.Path;
 
 public interface Dataset<Dataset_> extends Comparable<Dataset_> {
@@ -8,7 +9,8 @@ public interface Dataset<Dataset_> extends Comparable<Dataset_> {
 
     int ordinal();
 
-    double getBestKnownDistance();
+    // BigDecimal for precision.
+    BigDecimal getBestKnownDistance();
 
     boolean isLarge();
